+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 * [Flurry Android SDK Release Notes](https://developer.yahoo.com/flurry/docs/releasenotes/android/)
 * [Flurry iOS SDK Release Notes](https://developer.yahoo.com/flurry/docs/releasenotes/ios/)
 
+## v1.1.0 (2021-12-06)
+
+#### Features
+
+* Upgrade Flurry Android SDK version to 13.1.0
+* Upgrade Flurry iOS SDK version to 11.4.0
+* Support Agent setters API
+* Support Standard Event API
+* Support Publisher Segmentation API
+
 ## 0.1.0 (2021-09-30)
 
 * Create flutter-flurry-sdk package